+++ src/lib.rs
@@ -6,6 +6,7 @@ pub mod archive;
 pub mod audio;
 pub mod backend;
 pub mod bidi;
+pub mod cache;
 #[cfg(any(test, feature = "demo"))]
 pub mod demo;
 pub mod emoji;